move struct_lint_level to levels.rs
